requirements.txt

gradio


sentence-transformers

torch


newspaper3k


google-search-results


nltk